[George & Alex] saving functionality in sharing electoral info page
